backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg[binary]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
python-dotenv
httpx
